Church Celestial (#1006)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

This PR adds the ability to speak the Celestial language to adult Church
roles. This includes the priest, acolytes, templars, clerics and
paladins. Does not include churchlings or the Inquisition.

Incidentally fixes a formatting quirk in templar.dm.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game

Celestial currently has very little use in the game with it only being
spoken by the divine Aasimar, and even the Aasimar rarely get the
opportunity to use it in roleplay with the church. Having educated
religious roles be able to use the language would give it a tangible
ingame presence as a divine language to contrast with Infernal and Old
Psydonic.

Some reasons why it's a good idea:
- Clergy can bless (or curse) someone in divine scripture.
- Prompts roleplay between Aasimar and Church characters.
- Church roles can speak in a holy tongue when invoking the gods.
- The Church of the Ten have their own language separate from the
Psydonians and the Inhumens.
- Gives the Celestial language more flavour.
- Praying to the Ten in Celestial is _extremely_ pious.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->