A tale of maths and money (#959)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
This PR adds: 
Intelligence, perception and speed buffs or debuffs skills in checks
when counting and splitting coin stacks to better reflect brain power in
basic arithmetic.
Adds a small description based on the amount of coins for people who
aren't skilled.
This PR removes:
The inherent ability to "know" how much coins are worth (now it's an
estimate with an error based on the character's math skill).
Removes the (1 to max coin) prompt when splitting stacks so people can't
use it to skip coin counting.
Replaces the roman numbers used to describe how many coins there are in
a stack (why was that even a thing ?).
Adds a small description to singular coins that says how much they are
worth individually.
 
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game
This finally gives a very very noticeable impact for counting coins to
the game which can add a massive amount of RP potential as in traders
scamming people, uneducated or low intelligence people just having a
harder time to use currency.

Picture a half orc barbarian going to the merchant to sell a statue, the
merchant with this is incentivized to use low denomination coins to try
to scam the Horc into taking a noticeable less amount of coin, other
people will be overpaying unless they use higher denomination coins.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

Co-authored-by: grungussuss <96586172+Sadboysuss@users.noreply.github.com>