Dredge Adventuring Class (#984)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

Adds the Dredge!

The Dredge is a randomized adventurer **capped at 7 slots**. On spawn it
picks between a random armor type with associated trait, a weapon type
with skills and associated weapon, a random 'job' which gives the player
minor bonuses in a skill, and a random 'perk' which could give you
anything ranging from wine and hardtack, to a critical weakness flaw.

The skills and stats are **HIGHLY VARIABLE.** I've tried to minimize the
chances to roll for weaponskills above skilled, but it can happen, it's
the nature of RNG. There's a whole lot to cover, but predominantly there
are no steel weapons and the rarest 'armor' roll doesn't have full steel
armor.

## Why It's Good For The Game

People like to gamble. Don't want to play a Mage or a Thief? Have Dredge
as a pick? Roll the dice, you'll probably suffer, but you have a
horrendously low chance to spawn with a gun. The world is your oyster.

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->