Decorated Bascinet Recipe Fix (#1072)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

Fixes Decorated Bascinet taking 1 steel extra in cost, putting it in
line with other decorated helmet's cost.

## Why It's Good For The Game

One rarely used decorated helmet should not be costing 2 steel, may
aswell just get a knight helmet at that point.

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->