Hilly Murderwoods + Bandit Den Fix (#1077)
## About The Pull Request

The murderwoods had very flat hills--now they do not. There are roofed
caves and rocks on the upper Z-levels to mine. The bandit's den was also
not shifted properly on all z-levels, so now the tower's upper levels
are fixed. The roofs are also dilapidated from wear and so grass has
been growing under the holes.

## Why It's Good For The Game

It is a simple QOL change for the forest, and a necessary bug fix for
the bandit's den.

## Pre-Merge Checklist
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.