No one knows a foreigner's job (#1123)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

![image](https://github.com/user-attachments/assets/cb9d4a01-d2ea-4f42-9093-7ce6e2e39750)

![image](https://github.com/user-attachments/assets/27efda09-9f45-47ad-9739-37c5dc46a9e1)

![image](https://github.com/user-attachments/assets/6adafa1e-7494-4d97-b80d-652036331b17)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game
This was rolled back because there was a problem with pilgrim jobs not
being displayed, then I also rolled back all pilgrims being foreigners
so I think it makes sense again here.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [ ] You tested this on a local server.
- [ ] This code did not runtime during testing.
- [ ] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->