Merchant List Update + Brigandine Edit (#1078)
Updates merchant list pricing to be more logical, following the in-game
economy setup by players and the prices we typically see from Smiths.
Alongside pricing that makes sense for the amount of work that goes into
any item, and it's actual mechanical values.

Examples are Armor being priced slightly higher then what Smiths will
charge, but still alongside the general idea of ingot and ore values.

Health potions being priced higher so that the Wizard, Apothecary and
Feldscher's skills in alchemy are not devalued. This is also important
for ensuring that magical potions in a dark fantasy setting, are not
outright seen as commonplace mundane items. In future I will be doing
some extensive work on alchemy and potions once the Maintainers are done
with their fancy overhauls.

The merchant alongside these price edits has more items to sell,
including poisons, more armors, more weapons and better quality in terms
of what they have on offer.

Updated the Brigandine also in Define and the value for brig/coat of
plates. For history nerds, the Brigandine is actually less protective
then a Coat of Plates since it uses smaller overlapping plates as
opposed to the Coats larger more durable ones. So now the Brigandine has
lower protection, but higher integrity then a Coat of Plates.

Updated the sales price of riddle of steel so that it is more often sold
to the Court or used for it's actual purposes. It's buy price in the
merchant vendor is also 1100 from 1500. 1500 mammon is basically the
entire court's budget which is excessive. I still highly doubt anyone
will ever buy this.

Also this has been tested locally twice and is functional.