S&S sprite modification. (#1184)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
Batch of somewhat related sprite updates. 
Old vs. New:

![image](https://github.com/user-attachments/assets/d42e8bd8-4ad2-4961-ba40-aff507998132)

![image](https://github.com/user-attachments/assets/1be0e3b1-7fc6-446e-a583-c059562bcdec)

![image](https://github.com/user-attachments/assets/b337b2f3-b52f-4ee0-b6bc-018f7cc32f6f)

Fixes crop issues and miscoloured/misaligned trees.

![image](https://github.com/user-attachments/assets/cd53bd03-d654-4fc3-b18c-c98ca0d3a7a5)

Reverts the reversion of shards... for some reason??

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

Co-authored-by: AberrantQuerist <crystalseeker04@gmail.com>