Hidden Door Unmeta (#1012)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
Doors in the keep have a serious meta issue where people act as if they
always knew they were there, even if they really shouldn't. These doors
now require you to *learn* about them before you use them. Monarch,
Prince, Hand, and Butler know about them inherently. To everyone else,
too bad, figure it out yourself **[THIS EMOJI WAS STOPPING THE CODE FROM
COMPILING]** (or check the code if you want to ruin the fun). Also,
hidden doors have had smoothing implemented, so they will now be flush
with the rest of their respective walls.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->


## Why It's Good For The Game

The locks on the lord's door are currently a suggestion. This _sucks_
and I don't really feel like I have to say anything else.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

Co-authored-by: grungussuss <96586172+Sadboysuss@users.noreply.github.com>